src/main/java/ru/aafanasiev/util/copier/converter/clocaldate/CopyConverterLocalDateTime2OffsetTime.java
/**
 *
 */
package ru.aafanasiev.util.copier.converter.clocaldate;

import java.time.LocalDateTime;
import java.time.OffsetTime;
import java.time.ZoneId;

import ru.aafanasiev.annotations.Converter;
import ru.aafanasiev.util.copier.converter.base.ConverterKey;
import ru.aafanasiev.util.copier.converter.base.CopyConverter;
import ru.aafanasiev.util.copier.converter.base.CopyEntry;

/**
 * @author aafanasyev
 */
@Converter
public class CopyConverterLocalDateTime2OffsetTime implements CopyConverter {

    /*
     * (non-Javadoc)
     * @see ru.aafanasiev.util.copier.converter.base.CopyConverter#convert(ru.aafanasiev.util.copier.converter.base.
     * CopyEntry, java.lang.Object)
     */
    @Override
    public Object convert(CopyEntry entry, Object value) {
        if(null != value && LocalDateTime.class.equals(value.getClass())) {
            return ((LocalDateTime) value).atZone(ZoneId.systemDefault()).toOffsetDateTime().toOffsetTime();
        }
        return null;
    }

    /*
     * (non-Javadoc)
     * @see ru.aafanasiev.util.copier.converter.base.CopyConverter#listConvertedPairs()
     */
    @Override
    public ConverterKey[] listConvertedPairs() {
        return new ConverterKey[] {new ConverterKey(LocalDateTime.class, OffsetTime.class)};
    }

}
